Move utils back to tracing plugin as shared functionality
